pyproject.toml
[project]
name = "alcove"
version = "0.1.1"
description = "Add your description here"
authors = [{ name = "Lars Yencken", email = "lars@yencken.org" }]
dependencies = [
    "boto3>=1.34.145",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "jsonschema>=4.23.0",
    "ruff>=0.5.5",
    "rich>=13.7.1",
    "polars>=1.4.0",
    "duckdb>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
alcove = "alcove:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["pytest>=8.3.1", "minio>=7.2.7"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/alcove"]

[tool.uv]
dev-dependencies = [
    "minio>=7.2.10",
    "pyright>=1.1.387",
    "pytest>=8.3.3",
    "ruff>=0.7.2",
    "pytest-docker>=2.0.0",
    "docker>=7.0.0",
]
